requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-dateutil>=2.8.2
supabase>=1.0.3
python-dotenv>=1.0.0